+++ packages/core/src/web/index.ts
@@ -11,6 +11,7 @@ export * from "./musicLevelHistory";
 export * from "./openapi";
 export * from "./playCount";
 export * from "./playHistory";
+export * from "./progressionLinearRegression";
 export * from "./ratingAnalysis";
 export * from "./ratingBreakdownImageStorage";
 export * from "./ratingHistory";